+++ packages/traceloop-sdk/traceloop/sdk/client/client.py
@@ -1,6 +1,7 @@
 import sys
 
 from traceloop.sdk.annotation.user_feedback import UserFeedback
+from traceloop.sdk.datasets.datasets import Datasets
 from .http import HTTPClient
 from traceloop.sdk.version import __version__
 
@@ -38,3 +39,4 @@ class Client:
         self.api_key = api_key
         self._http = HTTPClient(base_url=self.api_endpoint, api_key=self.api_key, version=__version__)
         self.user_feedback = UserFeedback(self._http, self.app_name)
+        self.datasets = Datasets(self._http)
