--- packages/traceloop-sdk/traceloop/sdk/datasets/__init__.py
@@ -1,6 +1,0 @@
-from traceloop.sdk.datasets.dataset import Dataset
-from traceloop.sdk.datasets.column import Column
-from traceloop.sdk.datasets.row import Row
-from traceloop.sdk.datasets.model import ColumnType, DatasetMetadata
-
-__all__ = ["Dataset", "Column", "Row", "ColumnType", "DatasetMetadata"]
